requirements.txt
pytest
mysql-connector-python
Faker
requests
pytest-html
python-dotenv